+++ plugins/module_utils/utils/ansible_utils.py
@@ -65,8 +65,8 @@ def ansible_return(module, rsp, changed, req=None, existing_obj=None,
                 rsp.status_code, rsp.text, req, api_context))
     api_creds = AviCredentials()
     api_creds.update_from_ansible_module(module)
-    key = '%s:%s:%s' % (api_creds.controller, api_creds.username,
-                        api_creds.port)
+    key = '%s:%s:%s:%s' % (api_creds.controller, api_creds.username,
+                           api_creds.port, api_creds.tenant)
     deactivate_fact = module.params.get('avi_deactivate_session_cache_as_fact')
 
     fact_context = None
@@ -290,6 +290,7 @@ def avi_obj_cmp(x, y, sensitive_fields=None):
         y.pop('_last_modified', None)
         x.pop('api_version', None)
         y.pop('api_verison', None)
+        x.pop('verify', None)
         d_xks = [k for k in x.keys() if k in sensitive_fields]
 
         if d_xks:
@@ -350,8 +351,8 @@ def get_api_context(module, api_creds):
         return api_context
     elif api_context and not module.params.get(
             'avi_deactivate_session_cache_as_fact'):
-        key = '%s:%s:%s' % (api_creds.controller, api_creds.username,
-                            api_creds.port)
+        key = '%s:%s:%s:%s' % (api_creds.controller, api_creds.username,
+                               api_creds.port, api_creds.tenant)
         return api_context.get(key)
     else:
         return None
@@ -409,7 +410,9 @@ def avi_ansible_api(module, obj_type, sensitive_fields):
             token=api_context['csrftoken'],
             port=api_creds.port,
             session_id=api_context['session_id'],
-            csrftoken=api_context['csrftoken'])
+            csrftoken=api_context['csrftoken'],
+            verify=getattr(api_creds, 'verify', False),
+            avi_credentials=api_creds)
     else:
         api = ApiSession.get_session(
             api_creds.controller,
@@ -424,7 +427,9 @@ def avi_ansible_api(module, obj_type, sensitive_fields):
             csp_host=api_creds.csp_host,
             csp_token=api_creds.csp_token,
             ssl_cert=api_creds.ssl_cert,
-            ssl_key=api_creds.ssl_key)
+            ssl_key=api_creds.ssl_key,
+            verify=getattr(api_creds, 'verify', False),
+            avi_credentials=api_creds)
     state = module.params['state']
     # Get the api version.
     avi_update_method = module.params.get('avi_api_update_method', 'put')
@@ -467,7 +472,7 @@ def avi_ansible_api(module, obj_type, sensitive_fields):
         # As per API response, name is always same as username regardless of full_name
         obj['name'] = obj['username']
 
-    log.info('passed object %s ', obj)
+    log.info('passed object %s ', {**obj, 'password': None})
 
     if uuid:
         # Get the object based on uuid.
@@ -552,8 +557,15 @@ def avi_ansible_api(module, obj_type, sensitive_fields):
             obj_uuid = existing_obj['uuid']
             obj_path = '%s/%s' % (obj_type, obj_uuid)
         if avi_update_method == 'put':
-            changed = not avi_obj_cmp(obj, existing_obj, sensitive_fields)
-            obj = cleanup_absent_fields(obj)
+            # Merge with existing object to preserve unspecified fields
+            merged_obj = deepcopy(existing_obj)
+
+            for key, value in obj.items():
+                if value is not None:
+                    merged_obj[key] = value
+
+            changed = not avi_obj_cmp(merged_obj, existing_obj, sensitive_fields)
+            obj = cleanup_absent_fields(merged_obj)
             if changed:
                 req = obj
                 if check_mode:
@@ -619,7 +631,7 @@ def avi_common_argument_spec():
         controller=dict(default=os.environ.get('AVI_CONTROLLER', '')),
         username=dict(default=os.environ.get('AVI_USERNAME', '')),
         password=dict(default=os.environ.get('AVI_PASSWORD', ''), no_log=True),
-        api_version=dict(default=os.environ.get('API_VERSION', '')),
+        api_version=dict(default=os.environ.get('API_VERSION', '20.1.7')),
         tenant=dict(default='admin'),
         tenant_uuid=dict(default='', type='str'),
         port=dict(type='int'),
@@ -631,7 +643,8 @@ def avi_common_argument_spec():
         csp_host=dict(default='', type='str', no_log=True),
         csp_token=dict(default='', type='str', no_log=True),
         ssl_cert=dict(default='', type='str', no_log=True),
-        ssl_key=dict(default='', type='str', no_log=True)
+        ssl_key=dict(default='', type='str', no_log=True),
+        verify=dict(default=False)
     )
 
     return dict(
@@ -640,8 +653,9 @@ def avi_common_argument_spec():
         password=dict(default=os.environ.get('AVI_PASSWORD', ''), no_log=True),
         tenant=dict(default='admin'),
         tenant_uuid=dict(default=''),
-        api_version=dict(default=os.environ.get('API_VERSION', '')),
+        api_version=dict(default=os.environ.get('API_VERSION', '20.1.7')),
+        verify=dict(default=False),
         avi_credentials=dict(default=None, type='dict',
                              options=credentials_spec),
-        api_context=dict(type='dict'),
+        api_context=dict(type='dict', no_log=True),
         avi_deactivate_session_cache_as_fact=dict(default=False, type='bool'))