+++ plugins/modules/avi_sslkeyandcertificate.py
@@ -196,6 +196,15 @@ options:
             - Field introduced in 20.1.1.
             - Allowed with any value in enterprise, enterprise with cloud services edition.
         type: dict
+    skip_auto_chain:
+        description:
+            - When set to true, disables automatic ca certificate chain discovery based on issuer common name (cn).
+            - The user must explicitly specify the desired ca certificates via the ca_certs field.
+            - Not allowed for ca-type certificates.
+            - Field introduced in 32.1.3.
+            - Allowed with any value in enterprise, enterprise with cloud services edition.
+            - Default value when not specified in API or module is interpreted by Avi Controller as False.
+        type: bool
     status:
         description:
             - Enum options - SSL_CERTIFICATE_FINISHED, SSL_CERTIFICATE_PENDING.
@@ -310,6 +319,7 @@ def main():
         ocsp_error_status=dict(type='str',),
         ocsp_responder_url_list_from_certs=dict(type='list', elements='str',),
         ocsp_response_info=dict(type='dict',),
+        skip_auto_chain=dict(type='bool',),
         status=dict(type='str',),
         tenant_ref=dict(type='str',),
         type=dict(type='str',),
@@ -325,7 +335,7 @@ def main():
             'Python requests package is not installed. '
             'For installation instructions, visit https://pypi.org/project/requests.'))
     return avi_ansible_api(module, 'sslkeyandcertificate',
-                           {'key_passphrase', 'key', 'enckey_base64', 'key_params'})
+                           {'enckey_base64', 'key_passphrase', 'key_params', 'key'})
 
 
 if __name__ == '__main__':